Fix relay mode with new hardware.
The GPS software triggers a DTR and immediately sends
configuration data to the GPS, which puts the timer
into the bootloader. This adds a bypass to skip the
bootloader, while ensuring that we can always return
to the bootloader (by toggling DTR again).

timerutil --relay must now be run every time before
launching the GPS configuration software.